requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
